move http_client to utility
