+++ rlclientlib/rlclientlib.vcxproj
@@ -151,21 +151,21 @@
     <ClInclude Include="..\include\trace_logger.h" />
     <ClInclude Include="azure_factories.h" />
     <ClInclude Include="console_tracer.h" />
-    <ClInclude Include="http_client.h" />
     <ClInclude Include="logger\event_logger.h" />
     <ClInclude Include="moving_queue.h" />
     <ClInclude Include="utility\context_helper.h" />
     <ClInclude Include="utility\data_buffer.h" />
+    <ClInclude Include="utility\http_client.h" />
     <ClInclude Include="utility\interruptable_sleeper.h" />
     <ClInclude Include="utility\periodic_background_proc.h" />
     <ClInclude Include="utility\object_pool.h" />
     <ClCompile Include="async_batcher.cc" />
     <ClCompile Include="azure_factories.cc" />
     <ClCompile Include="console_tracer.cc" />
-    <ClCompile Include="http_client.cc" />
     <ClCompile Include="logger\event_logger.cc" />
     <ClCompile Include="trace_logger.cc" />
     <ClCompile Include="utility\data_buffer.cc" />
+    <ClCompile Include="utility\http_client.cc" />
     <ClCompile Include="utility\http_helper.cc" />
     <ClCompile Include="utility\http_helper.h" />
     <ClInclude Include="model_mgmt\model_downloader.h" />
