+++ rlclientlib/rlclientlib.vcxproj.filters
@@ -28,7 +28,7 @@
     <ClCompile Include="azure_factories.cc" />
     <ClCompile Include="async_batcher.cc" />
     <ClCompile Include="utility\http_helper.cc" />
-    <ClCompile Include="http_client.cc" />
+    <ClCompile Include="utility\http_client.cc" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="..\include\api_status.h" />
@@ -66,7 +66,7 @@
     <ClInclude Include="..\include\trace_logger.h" />
     <ClInclude Include="console_tracer.h" />
     <ClInclude Include="azure_factories.h" />
-    <ClInclude Include="http_client.h" />
+    <ClInclude Include="utility\http_client.h" />
   </ItemGroup>
   <ItemGroup>
     <None Include="packages.config" />
